added methods to identify server errors
